// src/components/EventContent.d.ts
import React from 'react';
import { ClassNames } from '../types';
export declare type EventContentProps = {
    width: number;
    height: number;
    classes: ClassNames;
    dateRange: [Date, Date];
    isStart: boolean;
    isEnd: boolean;
};
export declare const EventContent: React.NamedExoticComponent<EventContentProps>;
//# sourceMappingURL=EventContent.d.ts.map